- Adding NFD library to project itself. - Changing a few size_t types back to integers to remove warnings.
I don't know the legalities/ethical-ness of this, but as an amateur at C++, I would like to make the projects as easy as possible to download and run for other amateurs.
This includes having all necessary files available and pre-setup as much as possible. (Within reason. If I use a library that is too large or anything, then I won't do this.)